requirements.txt
fastapi==0.115.11
uvicorn==0.15.0
python-dotenv==0.19.0
pydantic==1.8.2
httpx==0.23.0
jinja2==3.0.1
aiofiles==0.7.0
websockets==10.0
anthropic==0.49.0
python-dotenv==1.0.1